type coversion change

+++ 1-Roots of Quadratic equation.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <stdlib.h>
 //calculating the roots of equation
-void calc_roots(int a, int b, int c)
+void calc_roots(double a, double b, double c)
 {
     if (a == 0)
     {
@@ -14,22 +14,22 @@ void calc_roots(int a, int b, int c)
     if (d > 0)
     {
         printf("Roots are both real and different \n");
-        printf("%f\n%f", (double)(-b + sqrt_val) / (2 * a), (double)(-b - sqrt_val) / (2 * a));
+        printf("%f\n%f", (-b + sqrt_val) / (2 * a), (-b - sqrt_val) / (2 * a));
     }
     else if (d == 0)
     {
         printf("Roots are real and same \n");
-        printf("%f", -(double)b / (2 * a));
+        printf("%f", -b / (2 * a));
     }
     else
     {
         printf("Roots are complex \n");
-        printf("%f + i%f\n%f - i%f", -(double)b / (2 * a), sqrt_val, -(double)b / (2 * a), sqrt_val);
+        printf("%f + i%f\n%f - i%f", -b / (2 * a), sqrt_val, -b / (2 * a), sqrt_val);
     }
 }
 int main()
 {
-    int a = 2, b = -5, c = 8;
+    double a = 2, b = -5, c = 8;
     calc_roots(a, b, c);
     return 0;
 }
